refactor memoization to use decorators
class method decorators, having access to the class instance which is
passed along with the function, are versatile and was implemented for
the memoization to eliminate redundant code.  even methods which take
different arguments can use the same decorator provided some branching
is done based on values of default arguments (or, if one prefers, on the
length of an args list when defining the inner_func to have a *args
input). speed is the same as before.